% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{atkins}
\alias{atkins}
\title{Assuming a linear model for the response, allocate treatment sequentially based on an optimal design criterion}
\usage{
atkins(covar, lossfunc = calc.D, int = NULL, init = 1,
  same.start = NULL, stoc = T, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{lossfunc}{a function for the optimality criterion to minimize}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{init}{the number of units in the initial design, by default set to 1}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}
}
\value{
design matrix and optimality at each sample size
}
\description{
Assuming a linear model for the response, allocate treatment sequentially based on an optimal design criterion
}
