% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{calc.A}
\alias{calc.A}
\title{Given a design matrix, calculate the A optimality for a linear model}
\usage{
calc.A(mat, epsilon = 1e-05)
}
\arguments{
\item{mat}{matrix or dataframe for the design matrix}

\item{epsilon}{a small real number used for regularization}
}
\value{
A optimality of the matrix
}
\description{
Given a design matrix, calculate the A optimality for a linear model
}
