% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{calc.bias}
\alias{calc.bias}
\title{Given a design, the current estimate of beta (or true value of beta), compute the bias matrix}
\usage{
calc.bias(beta, D, sim, true.bvcov = NULL)
}
\arguments{
\item{beta}{the current estimates of the parameter values}

\item{D}{current design matrix}

\item{sim}{number of simulated betas to generate}

\item{true.bvcov}{set to the true values of beta if the bias matrix is to be computed using the true values}
}
\value{
bias matrix
}
\description{
Given a design, the current estimate of beta (or true value of beta), compute the bias matrix
}
