% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{calc.DA}
\alias{calc.DA}
\title{Given a design matrix for a linear model, compute the D-A optimal objective function}
\usage{
calc.DA(mat, A = t(c(0, 0, 1, 0)), epsilon = 1e-05)
}
\arguments{
\item{mat}{matrix or dataframe for the design matrix}

\item{A}{a matrix where each column indicates the linear combination of parameters of interest}

\item{epsilon}{a small real number used for regularization}
}
\value{
value of the D-A optimal objective function
}
\description{
Given a design matrix for a linear model, compute the D-A optimal objective function
}
