% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{calc.Gint}
\alias{calc.Gint}
\title{Given a design matrix for a linear model wuth treatment-covariate interaction included, compute the G-optimal objective function}
\usage{
calc.Gint(mat, epsilon = 1e-05)
}
\arguments{
\item{mat}{matrix or dataframe for the design matrix}

\item{epsilon}{a small real number used for regularization}
}
\value{
value of the G optimal objective function
}
\description{
Given a design matrix for a linear model wuth treatment-covariate interaction included, compute the G-optimal objective function
}
