% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic.R
\name{calc.logit.wDA}
\alias{calc.logit.wDA}
\title{Calculate DA-optimal criterion assuming a logistic model (with or without weights)}
\usage{
calc.logit.wDA(Imat, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if weights equal}

\item{D}{design matrix}

\item{beta}{current estimate of coefficients}
}
\value{
DA-optimal criterion
}
\description{
Calculate DA-optimal criterion assuming a logistic model (with or without weights)
}
