% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic.R
\name{calc.logit.wL}
\alias{calc.logit.wL}
\title{Calculate L-optimal criterion assuming a logistic model (with or without weights)}
\usage{
calc.logit.wL(Imat, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if equal weights}

\item{D}{design matrix}

\item{beta}{current estimate of coefficients}
}
\value{
L-optimal criterion
}
\description{
Calculate L-optimal criterion assuming a logistic model (with or without weights)
}
