% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{calc.loss}
\alias{calc.loss}
\title{Given a design matrix for a linear model, compute the loss}
\usage{
calc.loss(mat, epsilon = 1e-05, int = NULL)
}
\arguments{
\item{mat}{matrix or dataframe for the design matrix}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}

\item{int}{set to NULL if there are no interactions, set to TRUE of there are interactions}
}
\value{
loss of the matrix
}
\description{
Given a design matrix for a linear model, compute the loss
}
