% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{calc.vcov}
\alias{calc.vcov}
\title{Given a design, the current estimate of beta (or true value of beta), compute the var-covar matrix}
\usage{
calc.vcov(beta, D, epsilon = 1e-05, true.bvcov = NULL)
}
\arguments{
\item{beta}{the current estimates of the parameter values}

\item{D}{current design matrix}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}

\item{true.bvcov}{set to the true values of beta if the var-covar matrix is to be computed using the true values}
}
\value{
var-covar matrix
}
\description{
Given a design, the current estimate of beta (or true value of beta), compute the var-covar matrix
}
