% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt linear.R
\name{calc.wDA}
\alias{calc.wDA}
\title{Calculate DA-optimal criterion assuming a linear model (with or without weights)}
\usage{
calc.wDA(D, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{D}{design matrix}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if weights equal}
}
\value{
DA-optimal criterion
}
\description{
Calculate DA-optimal criterion assuming a linear model (with or without weights)
}
