% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_linear.R
\name{calc.wL}
\alias{calc.wL}
\title{Compute the L-optimal objective function assuming a linear model (with or without weights)}
\usage{
calc.wL(D, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{D}{design matrix}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if equal weights}
}
\value{
Value of the L-optimal objective function
}
\description{
Compute the L-optimal objective function assuming a linear model (with or without weights)
}
