% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{calc.y.D}
\alias{calc.y.D}
\title{Given an information matrix and assuming logistic regression, compute the D-optimal objective function}
\usage{
calc.y.D(I, epsilon = 1e-05)
}
\arguments{
\item{I}{the information matrix}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}
}
\value{
value of the D-optimal objective function
}
\description{
Given an information matrix and assuming logistic regression, compute the D-optimal objective function
}
