% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{calc.y.DA}
\alias{calc.y.DA}
\title{Compute the DA-optimality criterion of the information matrix for logistic regression}
\usage{
calc.y.DA(I, A = t(matrix(c(0, 0, 1))), epsilon = 1e-05)
}
\arguments{
\item{I}{the information matrix}

\item{A}{a matrix where each column indicates the linear combination of parameters of interest}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}
}
\value{
DA-optimality of the information matrix
}
\description{
Compute the DA-optimality criterion of the information matrix for logistic regression
}
