% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt linear.R
\name{coord.cr}
\alias{coord.cr}
\title{Allocate treatment using a coordinate exchange algorithm with an optimality criterion for linear combinations of parameters.
Contrasts are assumed to be differences in treatment effect for each subgroup.}
\usage{
coord.cr(covar, k, cr, cr.lossfunc, prior.scale = 100)
}
\arguments{
\item{covar}{dataframe of covarite values}

\item{k}{integer for the number of "outer loop"}

\item{cr}{matrix of contrasts}

\item{cr.lossfunc}{a loss function appropriate for linear combinations of parameters.}

\item{prior.scale}{prior scale parameter}
}
\value{
Design matrix
}
\description{
Allocate treatment using a coordinate exchange algorithm with an optimality criterion for linear combinations of parameters.
Contrasts are assumed to be differences in treatment effect for each subgroup.
}
