% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_linear.R
\name{cr.des}
\alias{cr.des}
\title{Assuming a linear model for the response, allocate treatment sequentially based on an optimality criterion for
linear combinations of parameters. Responses are simulated assuming the true parameter values.}
\usage{
cr.des(covar, true.beta, true.sigma, threshold, kappa, init, cr.lossfunc, k, wt,
  int = T, prior.scale = 100, same.start = NULL, rand.start = NULL,
  stoc = T, bayes = T, prior.default = T, u = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the regression coefficients}

\item{true.sigma}{the true parameter value for the standard deviation}

\item{threshold}{the cut-off value for hypothesis tests}

\item{kappa}{the value of probability at which weights are set at zero}

\item{init}{the number of units in the initial design}

\item{cr.lossfunc}{loss function appropriate for linear combinations of parameters}

\item{k}{the number of "outer loops" in the coordinate exchange algorithm}

\item{wt}{set to T if the above lossfunction is weighted, NULL otherwise}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{prior.scale}{the prior scale parameter}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{prior.default}{set to T if default priors for bayesglm is used. If set to False and bayes=T, normal priors used.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
design matrix D, responses y, all estimates of betas, final estimate of beta, all weights, all estimates of standard deviation,
 beta, probabilities for treatment assignment, all values of optimalities (weighted L, DA, weighted DA), proportion of treatment=1,
 proportion of covariate in each group
 Type 1 error, true value of power, empirical value of power.
}
\description{
Assuming a linear model for the response, allocate treatment sequentially based on an optimality criterion for
linear combinations of parameters. Responses are simulated assuming the true parameter values.
}
