% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_nonmyop.R
\name{cr.exp.loss}
\alias{cr.exp.loss}
\title{Break down expected future optimality into two components:
1) assuming that the current response is 0
2) assuming that the current response is 1
Find the weighted average of the two cases}
\usage{
cr.exp.loss(z.now, t.now, zp, N, design, k, beta, y, cr, wr, prior.scale, bayes,
  dyn = NULL, Nprobs = NULL, ...)
}
\arguments{
\item{z.now}{vector of covariate values for current unit}

\item{t.now}{treatment of current unit}

\item{zp}{vector of probabilities for each level of covariate z (needs to in the same order as all.z below)}

\item{N}{natural number greater than 0 for horizon}

\item{design}{design matrix constructed for all units up until the current unit}

\item{k}{number of covariates}

\item{beta}{vector of current estimates for regression coefficients}

\item{y}{vectir of responses generated up until current unit}

\item{cr}{matrix of contrasts}

\item{wr}{matrix of weights}

\item{prior.scale}{prior scale parameter}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed}

\item{dyn}{set to NULL of there are no dynamic covariates, set to T if there are dynamic covariates}

\item{Nprobs}{a counter to be used if there are dynamic covariates}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
expected value of objective function one step ahead in the future
}
\description{
Break down expected future optimality into two components:
1) assuming that the current response is 0
2) assuming that the current response is 1
Find the weighted average of the two cases
}
