% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_linear_pseudononmy.R
\name{cr.future}
\alias{cr.future}
\title{Calculate (weighted) L-optimal expected optimality for a given trajectory using sequential approach.
We assume a linear model for the response.}
\usage{
cr.future(D.fix, n.r, sim, z.probs, lossfunc, cr, prior.scale, wr, ...)
}
\arguments{
\item{D.fix}{Design matrix constructed using the true covariates in the experiment so far}

\item{n.r}{length of trajectory to be simulated}

\item{sim}{number of trajectories to simulate}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{lossfunc}{the objective function to minimize}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if equal weights}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
loss of the design matrix which includes trajectory
}
\description{
Calculate (weighted) L-optimal expected optimality for a given trajectory using sequential approach.
We assume a linear model for the response.
}
