% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt nonmyop.R
\name{cr.future.y}
\alias{cr.future.y}
\title{Assuming a current response, break down the expected future optimality by cases for every combination of:
1) future possible covariate
2) future possible treatment
Find a weighted average across all these cases}
\usage{
cr.future.y(y.now, z.now, t.now, z.probs, N, design, k, beta, y, cr, wr,
  prior.scale, bayes, dyn = NULL, Nprobs = NULL, ...)
}
\arguments{
\item{z.now}{vector of covariate values for current unit}

\item{t.now}{treatment of current unit}

\item{z.probs}{vector of probabilities for each level of covariate z (needs to in the same order as all.z below)}

\item{N}{natural number greater than 0 for horizon}

\item{design}{design matrix constructed for all units up until the current unit}

\item{k}{number of covariates}

\item{beta}{vector of current estimates for regression coefficients}

\item{y}{vectir of responses generated up until current unit}

\item{cr}{matrix of contrasts}

\item{wr}{matrix of weights}

\item{prior.scale}{prior scale parameter}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed}

\item{dyn}{set to NULL of there are no dynamic covariates, set to T if there are dynamic covariates}

\item{Nprobs}{a counter to be used if there are dynamic covariates}
}
\value{
expected optimality assuming a current response
}
\description{
Assuming a current response, break down the expected future optimality by cases for every combination of:
1) future possible covariate
2) future possible treatment
Find a weighted average across all these cases
}
