% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic.R
\name{cr.logit.coord}
\alias{cr.logit.coord}
\title{Assuming a logistic model for the response, allocate treatments using a coordinate exchange algorithm according
to an information matrix-based optimality criterion.
Responses are simulated assuming the true parameter values.}
\usage{
cr.logit.coord(covar, beta, threshold, kappa,
  cr.lossfunc = calc.logit.wL, k, wt = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{beta}{an estimate for the true values of beta}

\item{threshold}{the cut-off value for hypothesis tests}

\item{kappa}{the value of probability at which weights are set at zero}

\item{cr.lossfunc}{loss function appropriate for optimality criteria which are for linear combinations of parameters}

\item{k}{the number of "outer loops" in the coordinate exchange algorithm}

\item{wt}{set to T if the above lossfunction is weighted, NULL otherwise}
}
\value{
design matrix D
}
\description{
Assuming a logistic model for the response, allocate treatments using a coordinate exchange algorithm according
to an information matrix-based optimality criterion.
Responses are simulated assuming the true parameter values.
}
