% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic pseudononmy.R
\name{cr.simfuture.logis.cont}
\alias{cr.simfuture.logis.cont}
\title{Allocate treatments according to weighted L-optimal objective function allowing for a pseudo-nonmyopic approach.
We assume a logistic model for the response and continuous treatment.}
\usage{
cr.simfuture.logis.cont(covar, true.beta, threshold, kappa, init, sim,
  z.probs, k = 2, wt, prior.scale = 100, same.start = NULL,
  rand.start = NULL, bayes = T, u = NULL, prior.default = T,
  true.bvcov = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the data generating mechanism}

\item{threshold}{the cut-off value for hypothesis tests}

\item{kappa}{the value of probability at which weights are set at zero}

\item{init}{the number of units in the initial design}

\item{sim}{number of trajectories to simulate}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{k}{number of "outer loops" in the coordinate exchange algorithm}

\item{wt}{set to T if the above lossfunction is weighted, NULL otherwise}

\item{prior.scale}{the prior scale parameter}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{prior.default}{set to T if default priors for bayesglm is used. If set to False and bayes=T, normal priors used.}

\item{true.bvcov}{if set to T, use the true parameter values to compute obejctive function. If set to NULL, use estimated parameter values.}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}
}
\value{
Design matrix D, optimality
}
\description{
Allocate treatments according to weighted L-optimal objective function allowing for a pseudo-nonmyopic approach.
We assume a logistic model for the response and continuous treatment.
}
