% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic pseudononmy.R
\name{cr.simfuture.logis}
\alias{cr.simfuture.logis}
\title{Allocate treatments according to an information matrix based optimality criterion allowing for a pseudo-nonmyopic approach.
We assume a logistic model for the response.}
\usage{
cr.simfuture.logis(covar, true.beta, threshold, kappa, init, int,
  cr.lossfunc, k, wt = T, sim, z.probs, same.start = NULL,
  rand.start = NULL, stoc = T, bayes = T, u = NULL,
  prior.default = T, coordex = NULL, true.bvcov = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the data generating mechanism}

\item{threshold}{the cut-off value for hypothesis tests}

\item{kappa}{the value of probability at which weights are set at zero}

\item{init}{the number of units in the initial design}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{cr.lossfunc}{a function for the optimality criterion to minimize}

\item{k}{number of "outer loops" in the coordinate exchange algorithm}

\item{wt}{set to T if the above lossfunction is weighted, NULL otherwise}

\item{sim}{number of trajectories to simulate}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{prior.default}{set to T if default priors for bayesglm is used. If set to False and bayes=T, normal priors used.}

\item{coordex}{set to T if the coordinate exchange algorithm is used to assign treatments in the trajectory. Sequential approach used otherwsise.}

\item{true.bvcov}{if set to T, use the true parameter values to compute obejctive function. If set to NULL, use estimated parameter values.}

\item{prior.scale}{the prior scale parameter}
}
\value{
Design matrix D, optimality
}
\description{
Allocate treatments according to an information matrix based optimality criterion allowing for a pseudo-nonmyopic approach.
We assume a logistic model for the response.
}
