% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{discrete}
\alias{discrete}
\title{discretize a vector of continuous values into binary (0 and 1) by using the}
\usage{
discrete(z)
}
\arguments{
\item{z}{a vector of continuous values}
}
\value{
a vector of descretized values
}
\description{
discretize a vector of continuous values into binary (0 and 1) by using the
}
