% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic pseudononmy.R
\name{Dopt.pseudo.t}
\alias{Dopt.pseudo.t}
\title{Calculate average D-optimal criterion assuming a logistic model with continuous treatment after M possible trajectories}
\usage{
Dopt.pseudo.t(t, D, z, int = NULL, beta, M, n.r, z.probs)
}
\arguments{
\item{t}{new treatment}

\item{D}{current deisgn matrix}

\item{z}{covariate values of new unit}

\item{int}{set to TRUE if treatment-covariate interactions are included in the model}

\item{beta}{current estimate of regression parameters}

\item{M}{number of trajectories}

\item{n.r}{length of each trajectory}

\item{z.probs}{assumed covariate distribution}
}
\value{
average value of D-optimal objective function across M trajectories
}
\description{
Calculate average D-optimal criterion assuming a logistic model with continuous treatment after M possible trajectories
}
