% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{Dopt.y.t}
\alias{Dopt.y.t}
\title{Given a design matrix with an additional row, and assuming assuming logistic regression, compute the D-optimal objective function
(to be used in conjunction with optim)}
\usage{
Dopt.y.t(t, D, z, int, beta, epsilon = 1e-05)
}
\arguments{
\item{t}{new treatment}

\item{D}{current deisgn matrix}

\item{z}{covariate values of new unit}

\item{int}{set to TRUE if treatment-covariate interactions are included in the model}

\item{beta}{current estimate of regression parameters}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}
}
\value{
value of the D-optimal objective function
}
\description{
Given a design matrix with an additional row, and assuming assuming logistic regression, compute the D-optimal objective function
(to be used in conjunction with optim)
}
