% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_nonmyop.R
\name{exp.loss.k}
\alias{exp.loss.k}
\title{Break down the expected future objective function by cases for every combination of:
1) future possible covariate
2) future possible treatment
Find a weighted average across all these cases}
\usage{
\method{exp}{loss.k}(z.now, t.now, zp, N, design, int, lossfunc, dyn = NULL,
  ...)
}
\arguments{
\item{z.now}{vector of covariate values for current unit}

\item{t.now}{treatment of current unit}

\item{zp}{vector of probabilities for each level of covariate z (needs to in the same order as all.z)}

\item{N}{natural number greater than 0 for horizon}

\item{design}{design matrix constructed for all units up until the current unit}

\item{int}{set to NULL if there are no interactions, set to T of there are interactions}

\item{lossfunc}{the objective function to minimize}

\item{dyn}{set to NULL of there are no dynamic covariates, set to T if there are dynamic covariates}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
design matrix D
}
\description{
Break down the expected future objective function by cases for every combination of:
1) future possible covariate
2) future possible treatment
Find a weighted average across all these cases
}
