% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic pseudononmy.R
\name{future.coordex.logis}
\alias{future.coordex.logis}
\title{Calculate expected optimality for a given trajectory using the coordinate exchange approach
We assume a logistic model for the response.}
\usage{
future.coordex.logis(D.fix, n.r, n, z.probs, beta, k, int, sim,
  code = NULL, lossfunc, ...)
}
\arguments{
\item{D.fix}{Design matrix constructed using the true covariates in the experiment so far}

\item{n.r}{length of trajectory to be simulated}

\item{n}{total number of patients in the experiment}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{beta}{current estimate of regression coefficients}

\item{k}{number of "outer loops" in the coordinate exchange algorithm}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{sim}{number of trajectories to simulate}

\item{code}{set to NULL if (-1,1) coding is used for the treatments. Set to 0 if (0, 1) is used.}

\item{lossfunc}{a function for the optimality criterion to minimize}
}
\value{
loss of the design matrix which includes trajectory
}
\description{
Calculate expected optimality for a given trajectory using the coordinate exchange approach
We assume a logistic model for the response.
}
