% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudononmy.R
\name{future.coordex}
\alias{future.coordex}
\title{Calculates expected optimality for a given trajectory after allocating treatments to the trajectory with
the coordinate exchange algorithm}
\usage{
future.coordex(D.fix, n.r, n, k, sim, int, z.probs, code = NULL,
  lossfunc, ...)
}
\arguments{
\item{D.fix}{Design matrix constructed using the true covariates in the experiment so far}

\item{n.r}{length of trajectory to be simulated}

\item{n}{number of total patients}

\item{k}{number of "outer loops" in the coordinate exchange algorithm}

\item{sim}{number of trajectories to simulate}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{code}{set to NULL if (-1,1) coding is used for the treatments. Set to 0 if (0, 1) is used.}

\item{lossfunc}{a function for the optimality criterion to minimize}
}
\value{
loss of the design matrix which includes trajectory
}
\description{
Calculates expected optimality for a given trajectory after allocating treatments to the trajectory with
the coordinate exchange algorithm
}
