% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic pseudononmy.R
\name{future.logis.cont}
\alias{future.logis.cont}
\title{Calculate expected optimality for a given trajectory using sequential approach when the treatment is continuous
We assume a logistic model for the response.}
\usage{
future.logis.cont(D.fix, n.r, z.probs, beta, int, sim, lossfunc, ...)
}
\arguments{
\item{D.fix}{Design matrix constructed using the true covariates in the experiment so far}

\item{n.r}{length of trajectory to be simulated}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{beta}{current estimate of regression coefficients}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{sim}{number of trajectories to simulate}

\item{lossfunc}{a function for the optimality criterion to minimize}
}
\value{
loss of the design matrix which includes trajectory
}
\description{
Calculate expected optimality for a given trajectory using sequential approach when the treatment is continuous
We assume a logistic model for the response.
}
