% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear nonmyop.R
\name{future.loss.k}
\alias{future.loss.k}
\title{Assuming the currrent response, future covariate value and future treatment,
Calculate optimality if horizon is 1. If not, iterate back to exp.loss function.}
\usage{
future.loss.k(z.next, t.next, z.now, t.now, z.probs, N, design, int,
  lossfunc, dyn = NULL, Nprobs = NULL, ...)
}
\arguments{
\item{z.next}{vector of covariate values for future unit}

\item{t.next}{treatment of future unit}

\item{z.now}{vector of covariate values for current unit}

\item{t.now}{treatment of current unit}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{N}{natural number greater than 0 for horizon}

\item{design}{design matrix constructed for all units up until the current unit}

\item{int}{set to NULL if there are no interactions, set to T of there are interactions}

\item{lossfunc}{a function for the optimality criterion to minimize}

\item{dyn}{set to NULL of there are no dynamic covariates, set to T if there are dynamic covariates}

\item{Nprobs}{a counter to be used if there are dynamic covariates}
}
\value{
optimality one step ahead in the future, assuming z.next and t.next
}
\description{
Assuming the currrent response, future covariate value and future treatment,
Calculate optimality if horizon is 1. If not, iterate back to exp.loss function.
}
