% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic nonmyopic.R
\name{future.loss}
\alias{future.loss}
\title{Assuming the currrent response, future covariate value and future treatment,
Calculate optimality if horizon is 1. If not, iterate back to exp.loss function.}
\usage{
future.loss(z.next, t.next, zp, N, design, int, lossfunc, beta, y, bayes,
  dyn = NULL, ...)
}
\arguments{
\item{z.next}{vector of covariate values for future unit}

\item{t.next}{treatment of future unit}

\item{zp}{vector of probabilities for each level of covariate z (needs to in the same order as all.z below)}

\item{N}{natural number greater than 0 for horizon}

\item{design}{design matrix constructed for all units up until the current unit}

\item{int}{set to NULL if there are no interactions, set to T of there are interactions}

\item{lossfunc}{a function for the optimality criterion to minimize}

\item{beta}{estimate of the regression coefficients}

\item{y}{responses that have been observed up until the current unit}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{z.now}{vector of covariate values for current unit}

\item{t.now}{treatment of current unit}
}
\value{
optimality assuming current response, future covariate value and future treatment
}
\description{
Assuming the currrent response, future covariate value and future treatment,
Calculate optimality if horizon is 1. If not, iterate back to exp.loss function.
}
