% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gencov}
\alias{gencov}
\title{Create a set of binary covariates with a specific covariate distribution}
\usage{
gencov(z.probs, n.r, code = NULL)
}
\arguments{
\item{z.probs}{if a scalar, assume there is one binary covariate z and
the scalar is P(z=1).
if z.probs is a vector of length k, there are k static covariates, z1, ..., zk, and
the the jth element of the vector is  P(zj=1).
if z.probs is a matrix with k columns and n.r rows, there are k dynamic covariates z1, ...zk.
the (i,j)th element is the probability that the jt unit of the ith covariate is equal to 1}

\item{n.r}{number of covariate values to generate}

\item{code}{set to NULL if (-1,1) coding is used for the treatments. Set to 0 if (0, 1) is used.}

\item{mat}{matrix or dataframe for the design matrix}
}
\value{
dataframe with covariates
}
\description{
Create a set of binary covariates with a specific covariate distribution
}
