% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{Imat.beta}
\alias{Imat.beta}
\title{Compute the information matrix, given beta, for logistic regression}
\usage{
Imat.beta(X, beta)
}
\arguments{
\item{X}{the design matrix}

\item{beta}{coefficients of the model}
}
\value{
Information matrix
}
\description{
Compute the information matrix, given beta, for logistic regression
}
