% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_nonmyop.R
\name{learn.zprobs}
\alias{learn.zprobs}
\title{Find the empirical distribution of the covariates}
\usage{
learn.zprobs(D, all.z, j)
}
\arguments{
\item{D}{design matrix}

\item{all.z}{dataframe containing all covariate combinations}

\item{j}{number of covariates}
}
\value{
dataframe z.probs with distribution of covariates
}
\description{
Find the empirical distribution of the covariates
}
