% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear nonmyop.R
\name{linear.nonmyop.dyn}
\alias{linear.nonmyop.dyn}
\title{Allocate treatments according to an information matrix based optimality criterion allowing for a non-myopic approach.
Allow for dynamic covariates. We assume a linear model for the response and simulate responses sequentially.}
\usage{
linear.nonmyop.dyn(covar, init, z.probs, k, N, int = NULL,
  lossfunc = calc.D, stoc = F, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{init}{the number of units in the initial design}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{k}{integer for number of "outer" loops in coordinate exchange algorithm for initial design}

\item{N}{natural number greater than 0 for horizon}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{lossfunc}{a function for the optimality criterion to minimize}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}
}
\value{
Design matrix D
}
\description{
Allocate treatments according to an information matrix based optimality criterion allowing for a non-myopic approach.
Allow for dynamic covariates. We assume a linear model for the response and simulate responses sequentially.
}
