% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logit.coord}
\alias{logit.coord}
\title{Assuming a logistic model for the response, allocate treatments using a coordinate exchange algorithm according
to an information matrix-based optimality criterion.
Responses are simulated assuming the true parameter values.}
\usage{
logit.coord(covar, beta, k, int = NULL, code = NULL, lossfunc = calc.y.D,
  ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{beta}{an estimate for the true values of beta}

\item{k}{an integer for the number of "outer loops"}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{code}{set to NULL if (-1,1) coding is used for the treatments. Set to 0 if (0, 1) is used.}

\item{lossfunc}{the objective function to minimize}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
design matrix D
}
\description{
Assuming a logistic model for the response, allocate treatments using a coordinate exchange algorithm according
to an information matrix-based optimality criterion.
Responses are simulated assuming the true parameter values.
}
