% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{logit.mse}
\alias{logit.mse}
\title{Allocate treatments according to the MSE matrix when a logisic model for the response is assumed.
We simulate responses sequentially.}
\usage{
logit.mse(covar, true.beta, init, int = NULL, sim, lossfunc = calc.y.D,
  epsilon = 1e-05, same.start = NULL, true.bvcov = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the data generating mechanism}

\item{init}{the number of units in the initial design}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{lossfunc}{a function for the optimality criterion to minimize}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}

\item{same.start}{set to the intial design if desired or set to NULL otherwise #'}

\item{true.bvcov}{set to the true values of beta if the mse matrix is to be computed using the true values}
}
\value{
the design matrix D, responses y, all estimates of beta, final estimate of beta, probabilities of treatment
assignment, proportion of favorable responses, optimality, trace of var-covar matrix, trace of bias matrix,
trace of mse matrix, determinant of mse matrix
}
\description{
Allocate treatments according to the MSE matrix when a logisic model for the response is assumed.
We simulate responses sequentially.
}
