% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_nonmyopic.R
\name{logit.nonmy}
\alias{logit.nonmy}
\title{Allocate treatments according to an information matrix based optimality criterion allowing for a non-myopic approach.
We assume a logistic model for the response and simulate responses sequentially.}
\usage{
logit.nonmy(covar, true.beta, init, z.probs, N, int = NULL,
  lossfunc = calc.y.D, same.start = NULL, rand.start = NULL, stoc = T,
  bayes = T, u = NULL, true.bvcov = NULL, dyn = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the data generating mechanism}

\item{init}{the number of units in the initial design}

\item{z.probs}{probabilities for each covariate value being 1}

\item{N}{natural number greater than 0 for horizon}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{lossfunc}{the objective function to minimize}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{true.bvcov}{if set to T, use the true parameter values to compute obejctive function. If set to NULL, use estimated parameter values.}

\item{dyn}{set to T if there is a dynamic covariate}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
Design matrix D, all estimates of beta, final estimate of beta, responses y
}
\description{
Allocate treatments according to an information matrix based optimality criterion allowing for a non-myopic approach.
We assume a logistic model for the response and simulate responses sequentially.
}
