% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logit.rand}
\alias{logit.rand}
\title{Assuming a logistic model for the response, allocate treatment at random.
Responses are simulated assuming the true parameter values.}
\usage{
logit.rand(covar, true.beta, init, int = NULL, lossfunc = calc.y.D,
  same.start = NULL, rand.start = NULL, bayes = T, u = NULL,
  true.bvcov = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the data generating mechanism}

\item{init}{the number of units in the initial design}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{lossfunc}{the objective function to minimize}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{true.bvcov}{if set to T, use the true parameter values to compute obejctive function. If set to NULL, use estimated parameter values.}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
design matrix D, responses y, all estimates of betas, final beta, probabilities for treatment assignment,
proportion of Y=1, proportion tmt=1, values of objective function
}
\description{
Assuming a logistic model for the response, allocate treatment at random.
Responses are simulated assuming the true parameter values.
}
