% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{max.imb}
\alias{max.imb}
\title{Allocate treatment based on Hu and Hu (2012)'s version of minimization.
Hu and Hu (2012) imbalance measure (only applicable for one continuous covariate)
Compute imbalance measure proposed in Hu and Hu (2012):
sort list of currently assigned treatments according to the covariate
consider all subsets containing the current patient k
Find difference in number of patients who have been given treatment 1 vs treatment 0
Find the maximum difference across all subsets}
\usage{
\method{max}{imb}(T01, k)
}
\arguments{
\item{T01}{vector of assigned treatments, sorted according to the covariate}

\item{k}{position of the current patient in the sorted list T01}
}
\value{
max.imb
}
\description{
Allocate treatment based on Hu and Hu (2012)'s version of minimization.
Hu and Hu (2012) imbalance measure (only applicable for one continuous covariate)
Compute imbalance measure proposed in Hu and Hu (2012):
sort list of currently assigned treatments according to the covariate
consider all subsets containing the current patient k
Find difference in number of patients who have been given treatment 1 vs treatment 0
Find the maximum difference across all subsets
}
