% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{min.classic}
\alias{min.classic}
\title{Classic version of Minimization, appropriate for binary covariates only}
\usage{
\method{min}{classic}(covar, p = 2/3)
}
\arguments{
\item{covar}{a dataframe for covariates  (must be binary)}

\item{p}{takes value between 0.5 and 1 and determines randomness in treatment assignment}
}
\value{
design matrix
}
\description{
Classic version of Minimization, appropriate for binary covariates only
}
