% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{min.cont}
\alias{min.cont}
\title{Allocate treatment using a version of minimization, appropriate for continuous covariates}
\usage{
\method{min}{cont}(covar, p = 2/3)
}
\arguments{
\item{covar}{a dataframe for covariates  (must be continuous)}

\item{p}{takes value between 0.5 and 1 and determines randomness in treatment assignment}
}
\value{
design matrix
}
\description{
Allocate treatment using a version of minimization, appropriate for continuous covariates
}
