% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{min.gen}
\alias{min.gen}
\title{Allocate treatment based on an imbalance criterion - most general form of minimization}
\usage{
\method{min}{gen}(covar, imb, p = 2/3)
}
\arguments{
\item{covar}{a dataframe for one covariate ONLY (not applicable for multiple covariates)}

\item{imb}{is a measure of imbalance. It can be set to "efron", "ks" or "max.imb"}

\item{p}{takes value between 0.5 and 1 and determines randomness in treatment assignment}
}
\value{
design matrix
}
\description{
Allocate treatment based on an imbalance criterion - most general form of minimization
}
