% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{min.sen}
\alias{min.sen}
\title{Allocate treatment using Senn's version of Minimization, appropriate for binary covariates only.}
\usage{
\method{min}{sen}(covar, p = 2/3)
}
\arguments{
\item{covar}{a dataframe for covariates  (must be binary)}

\item{p}{takes value between 0.5 and 1 and determines randomness in treatment assignment}
}
\value{
design matrix
}
\description{
Allocate treatment using Senn's version of Minimization, appropriate for binary covariates only.
}
