% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{probi}
\alias{probi}
\title{Compute P(y=1) for logistic regression}
\usage{
probi(x, beta)
}
\arguments{
\item{x}{a row of the design matrix}

\item{beta}{coefficients of the model}
}
\value{
Probability P(y=1)'
}
\description{
Compute P(y=1) for logistic regression
}
