% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimization.R
\name{rand}
\alias{rand}
\title{Randomly allocate treatment}
\usage{
rand(covar)
}
\arguments{
\item{covar}{a dataframe for the covariates}
}
\value{
design matrix
}
\description{
Randomly allocate treatment
}
