% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{shadeplot}
\alias{shadeplot}
\title{Plot distribution}
\usage{
shadeplot(data, n1 = 4, n = 100, xlabel = "", ylabel = "", ...)
}
\arguments{
\item{data}{a dataframe with each column being the result of one experiment}

\item{n1}{the smallest sample size to display on the graph}

\item{n}{the largest sample size to display on the graph}

\item{xlabel}{the x axis label}

\item{ylabel}{the y axis label}
}
\value{
None
}
\description{
Creates a plot the distribution of a performance measure across sample size
}
\examples{
shadeplot(data, 10, 100, "Sample size", "D-optimality")

}
