% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_pseudononmy.R
\name{simfuture.logis.cont}
\alias{simfuture.logis.cont}
\title{Allocate continuous treatments according to an information matrix based optimality criterion allowing for a pseudo-nonmyopic approach.
We assume a logistic model for the response.}
\usage{
simfuture.logis.cont(covar, true.beta, init, k, sim, z.probs, int = NULL,
  lossfunc, same.start = NULL, rand.start = NULL, bayes = T, u = NULL,
  true.bvcov = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the data generating mechanism}

\item{init}{the number of units in the initial design}

\item{k}{number of "outer loops" in the coordinate exchange algorithm}

\item{sim}{number of trajectories to simulate}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{lossfunc}{the objective function to minimize}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{true.bvcov}{if set to T, use the true parameter values to compute obejctive function. If set to NULL, use estimated parameter values.}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
Design matrix D, value of the objective function
}
\description{
Allocate continuous treatments according to an information matrix based optimality criterion allowing for a pseudo-nonmyopic approach.
We assume a logistic model for the response.
}
