% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudononmy.R
\name{simfuture}
\alias{simfuture}
\title{Allocate treatments according to an information matrix based optimality criterion allowing for a pseudo-nonmyopic approach.
We assume a linear model for the response.}
\usage{
simfuture(covar, sim, int = NULL, z.probs, lossfunc, stoc = T,
  same.start = NULL, coordex = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{sim}{number of trajectories to simulate}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{lossfunc}{a function for the optimality criterion to minimize}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{coordex}{set to T if coordinate exchange algorithm is used to allocate treatments in the trajectory, set to NULL for sequential approach.}

\item{N}{natural number greater than 0 for horizon}
}
\value{
Design matrix D, optimality
}
\description{
Allocate treatments according to an information matrix based optimality criterion allowing for a pseudo-nonmyopic approach.
We assume a linear model for the response.
}
