% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic pseudononmy.R
\name{wLopt.pseudo.t}
\alias{wLopt.pseudo.t}
\title{Calculate average L-optimal criterion assuming a logistic model with continuous treatment after M possible trajectories}
\usage{
wLopt.pseudo.t(t, D, z, beta, cr, prior.scale = 100, wr = NULL, M, n.r,
  z.probs)
}
\arguments{
\item{t}{new treatment}

\item{D}{current deisgn matrix}

\item{z}{covariate values of new unit}

\item{beta}{current estimate of regression parameters}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if equal weights}

\item{M}{number of trajectories}

\item{n.r}{length of each trajectory}

\item{z.probs}{assumed covariate distribution}
}
\value{
average value of L-optimal objective function across M trajectories
}
\description{
Calculate average L-optimal criterion assuming a logistic model with continuous treatment after M possible trajectories
}
