% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_logistic.R
\name{wLopt.t}
\alias{wLopt.t}
\title{Compute the L-optimal objective function assuming a logistic model (with or without weights) assuming continuous treatment,
given one new patient}
\usage{
wLopt.t(t, D, z, beta, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{t}{new treatment}

\item{D}{current deisgn matrix}

\item{z}{covariate values of new unit}

\item{beta}{current estimate of coefficients}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if equal weights}
}
\value{
Value of the L-optimal objective function
}
\description{
Compute the L-optimal objective function assuming a logistic model (with or without weights) assuming continuous treatment,
given one new patient
}
