% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{zpfunc}
\alias{zpfunc}
\title{Given z.probs which takes a specific form, convert it to an nxp dataframe where the (i,j)th entry is the probability
that the ith patient's jth covariate value is 1}
\usage{
zpfunc(z.probs)
}
\arguments{
\item{z.probs}{if a scalar, assume there is one binary covariate z and
the scalar is P(z=1).
if z.probs is a vector of length k, there are k static covariates, z1, ..., zk, and
the the jth element of the vector is  P(zj=1).
if z.probs is a matrix with k columns and n.r rows, there are k dynamic covariates z1, ...zk.
the (i,j)th element is the probability that the jt unit of the ith covariate is equal to 1}
}
\value{
nxp dataframe where the (i,j)th entry is the probability
that the ith patient's jth covariate value is 1
}
\description{
Given z.probs which takes a specific form, convert it to an nxp dataframe where the (i,j)th entry is the probability
that the ith patient's jth covariate value is 1
}
