% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{calc.G}
\alias{calc.G}
\title{Given a design matrix, calculate the G optimality for a linear model}
\usage{
calc.G(mat, epsilon = 1e-05)
}
\arguments{
\item{mat}{matrix or dataframe for the design matrix}

\item{epsilon}{a small real number used for regularization}
}
\value{
G optimality of the matrix
}
\description{
Given a design matrix, calculate the G optimality for a linear model
}
