% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{calc.tz}
\alias{calc.tz}
\title{Given a design matrix, calculate the covariate balance for a linear model}
\usage{
calc.tz(mat)
}
\arguments{
\item{mat}{matrix or dataframe for the design matrix}
}
\value{
covariate balance of the matrix
}
\description{
Given a design matrix, calculate the covariate balance for a linear model
}
