% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_linear.R
\name{calc.wDA}
\alias{calc.wDA}
\title{Compute the DA-optimal objective function assuming a linear model (with or without weights)}
\usage{
calc.wDA(D, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{D}{design matrix}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if weights equal}
}
\value{
Value of the DA-optimal objective function
}
\description{
Compute the DA-optimal objective function assuming a linear model (with or without weights)
}
