% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{calc.y.D}
\alias{calc.y.D}
\title{Compute the D-optimality criterion of the information matrix for logistic regression}
\usage{
calc.y.D(I, epsilon = 1e-05)
}
\arguments{
\item{I}{the information matrix}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}
}
\value{
D-optimality of the information matrix
}
\description{
Compute the D-optimality criterion of the information matrix for logistic regression
}
