% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{calc.y.DA}
\alias{calc.y.DA}
\title{Given an information matrix and assuming logistic regression, compute the D-A optimal objective function}
\usage{
calc.y.DA(I, A = t(matrix(c(0, 0, 1))), epsilon = 1e-05)
}
\arguments{
\item{I}{the information matrix}

\item{A}{a matrix where each column indicates the linear combination of parameters of interest}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}
}
\value{
value of the D-A optimal objective function
}
\description{
Given an information matrix and assuming logistic regression, compute the D-A optimal objective function
}
