% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{coordex}
\alias{coordex}
\title{Allocate treatment using coordinate exchange algorithm, assuming a linear model for the response
which has an intercept, an effect for each covariate, an effect for the treatment, and allows for
covariate-treatment interactions}
\usage{
coordex(covar, k, lossfunc, int = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{k}{the number of random starting designs to use}

\item{lossfunc}{the objective function to minimize}

\item{int}{set to T if you allow for treatment-covariate interactions in the model, NULL otherwise}

\item{...}{Further arguments to be passed to <lossfunc>}
}
\value{
design matrix
}
\description{
Allocate treatment using coordinate exchange algorithm, assuming a linear model for the response
which has an intercept, an effect for each covariate, an effect for the treatment, and allows for
covariate-treatment interactions
}
