% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt_nonmyop.R
\name{cr.future.loss}
\alias{cr.future.loss}
\title{Assuming the currrent response, future covariate value and future treatment,
Calculate objective function if horizon is 1. If horizon >1, iterate back to cr.exp.loss function.}
\usage{
cr.future.loss(z.next, t.next, zp, N, design, k, beta, y, cr, wr, prior.scale,
  bayes, dyn = NULL, Nprobs = NULL, ...)
}
\arguments{
\item{z.next}{vector of covariate values for future unit}

\item{t.next}{treatment of future unit}

\item{zp}{vector of probabilities for each level of covariate z (needs to in the same order as all.z below)}

\item{N}{natural number greater than 0 for horizon}

\item{design}{design matrix constructed for all units up until the current unit}

\item{k}{number of covariates}

\item{beta}{vector of current estimates for regression coefficients}

\item{y}{vectir of responses generated up until current unit}

\item{cr}{matrix of contrasts}

\item{wr}{matrix of weights}

\item{prior.scale}{prior scale parameter}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed}

\item{dyn}{set to NULL of there are no dynamic covariates, set to T if there are dynamic covariates}

\item{Nprobs}{a counter to be used if there are dynamic covariates}

\item{...}{further arguments to be passed to <lossfunc>}
}
\value{
value of objective function assuming current response, future covariate value and future treatment
}
\description{
Assuming the currrent response, future covariate value and future treatment,
Calculate objective function if horizon is 1. If horizon >1, iterate back to cr.exp.loss function.
}
