% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{Dopt.y.t.init}
\alias{Dopt.y.t.init}
\title{Compute the D-optimality criterion of of an initial design for logistic regression
(to be used in conjunction with optim)}
\usage{
Dopt.y.t.init(t, z, int, beta, epsilon = 1e-05)
}
\arguments{
\item{t}{vector of treatments}

\item{int}{set to TRUE if treatment-covariate interactions are included in the model}

\item{beta}{current estimate of regression parameters}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}

\item{vector}{of covariate values}
}
\value{
D-optimality criterion
}
\description{
Compute the D-optimality criterion of of an initial design for logistic regression
(to be used in conjunction with optim)
}
