% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{Dopt.y.t}
\alias{Dopt.y.t}
\title{Compute the D-optimality criterion of the information matrix for logistic regression with an added row of the design matrix
(to be used in conjunction with optim)}
\usage{
Dopt.y.t(t, D, z, int, beta, epsilon = 1e-05)
}
\arguments{
\item{t}{new treatment}

\item{D}{current deisgn matrix}

\item{z}{covariate values of new unit}

\item{int}{set to TRUE if treatment-covariate interactions are included in the model}

\item{beta}{current estimate of regression parameters}

\item{epsilon}{a small real number used for regularization. If set to zero,
no regularization takes place}
}
\value{
D-optimality criterion
}
\description{
Compute the D-optimality criterion of the information matrix for logistic regression with an added row of the design matrix
(to be used in conjunction with optim)
}
