% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt nonmyop.R
\name{logit.Lbnon}
\alias{logit.Lbnon}
\title{Allocate treatments according to a weighted L-optimal criterion allowing for a non-myopic approach.
We assume a logistic model for the response and simulate responses sequentially.}
\usage{
logit.Lbnon(covar, true.beta, threshold, kappa, init, z.probs, N,
  prior.scale = 100, same.start = NULL, rand.start = NULL,
  stoc = T, bayes = T, u = NULL, true.bvcov = NULL, ...)
}
\arguments{
\item{covar}{a dataframe for the covariates}

\item{true.beta}{the true parameter values of the regression coefficients}

\item{threshold}{the cut-off value for hypothesis tests}

\item{kappa}{the value of probability at which weights are set at zero}

\item{init}{the number of units in the initial design}

\item{z.probs}{vector of probabilities for each level of covariate z}

\item{N}{natural number greater than 0 for horizon}

\item{prior.scale}{the prior scale parameter}

\item{same.start}{the design matrix to be used for the initial design. If set to NULL, function generates initial design.}

\item{rand.start}{If set to T, function generates an initial design randomly. Else, coordinate exchange is used.}

\item{stoc}{set to T if treatments are allocated using a stochastic method where the probability is
determined by the optimality crtierion. Set to F if treatments are allocated deterministically.}

\item{bayes}{set to T if bayesglm is used instead of glm. Default prior assumed.}

\item{u}{vector of uniform random numbers for generating responses. If set to NULL, responses generated from the binomial distribution.}

\item{true.bvcov}{if set to T, use the true parameter values to compute obejctive function. If set to NULL, use estimated parameter values.}
}
\value{
Design matrix D, all estimates of beta, final estimate of beta, responses y
}
\description{
Allocate treatments according to a weighted L-optimal criterion allowing for a non-myopic approach.
We assume a logistic model for the response and simulate responses sequentially.
}
