% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt logistic.R
\name{wLopt.t.init}
\alias{wLopt.t.init}
\title{Calculate L-optimal criterion assuming a logistic model (with or without weights) assuming continuous treatment,
given a vector of treatment values and a vector of covariates values for additional patients}
\usage{
wLopt.t.init(t, z, beta, cr, prior.scale = 100, wr = NULL)
}
\arguments{
\item{t}{new treatment}

\item{z}{covariate values of new unit}

\item{beta}{current estimate of coefficients}

\item{cr}{matrix of contrasts}

\item{prior.scale}{prior scale parameter}

\item{wr}{matrix of weights, set to NULL if equal weights}

\item{D}{design matrix}
}
\value{
L-optimal criterion
}
\description{
Calculate L-optimal criterion assuming a logistic model (with or without weights) assuming continuous treatment,
given a vector of treatment values and a vector of covariates values for additional patients
}
